backend/requirements.txt
# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.25.0

# Database
sqlalchemy>=2.0.0
alembic>=1.13.0

# Settings
pydantic-settings>=2.0.0

# Async Task Queue
huey>=2.5.0

# Scheduler
apscheduler>=3.10.0

# HTTP Client
httpx>=0.26.0

# RSS Parsing
feedparser>=6.0.0

# LLM Client (OpenAI compatible)
openai>=1.0.0

# Video Download
yt-dlp>=2024.1.0

# HTML Parsing
beautifulsoup4>=4.12.0
readability-lxml>=0.8.0
lxml>=5.0.0

# Finance Data
akshare>=1.14.0

# Utilities
python-multipart>=0.0.6
python-dotenv>=1.0.0
